@@ -4,35 +4,22 @@
   "description": "",
   "keywords": [],
   "main": "src/index.js",
-  "dependencies": {
-    "@testing-library/react": "9.4.0",
-    "@testing-library/react-hooks": "3.2.1",
-    "expect": "24.9.0",
-    "postcss": "7.0.21",
-    "prop-types": "15.7.2",
-    "react": "16.12.0",
-    "react-dom": "16.12.0",
-    "react-number-format": "4.3.1",
-    "react-query": "0.3.24",
-    "react-router-dom": "5.1.2",
-    "react-scripts": "3.3.0",
-    "react-spinners": "0.8.0",
-    "reselect": "4.0.0",
-    "tailwindcss": "1.1.4"
-  },
-  "devDependencies": {
-    "typescript": "3.3.3"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "dependencies": {
+    "prop-types": "15.8.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-number-format": "5.3.1",
+    "react-query": "3.39.3",
+    "react-router-dom": "^6.15.0",
+    "react-scripts": "^5.0.1",
+    "react-spinners": "0.13.8",
+    "reselect": "4.1.8",
+    "tailwindcss": "3.3.3"
+  }
 }
